internal/efi: abstract cpuid behind an arch-gated wrapper
github.com/canonical/cpuid does not compile on non-x86 architectures,
which forced the AMD64 host environment to be split across a filename
gated default_env_amd64.go and a default_env_amd64_null.go stub. That
made the implementation invisible to the compiler, and untestable, on
every other architecture.

Confine that dependency to a new internal/cpuid package. It is a
deliberately thin wrapper, and the only place in the tree that is gated
by architecture at build time. It exposes VendorIdentificator, Family
and HasFeature, delegating to the upstream package on amd64 and
returning zero values elsewhere.

internal/efi is now architecture neutral: it carries no build
constraints, the AMD64 implementation collapses into default_env.go, and
AMD64() selects on a mockable runtimeGOARCH rather than a build tag.
This mirrors the runtime dispatch used elsewhere for host security
checks.

The CPUIDFeature* constants stay in internal/efi, so callers are
unchanged. Their drift guards against the upstream bit positions move to
internal/cpuid, which is now the only package able to see both.

The AMD64 tests lose their build constraints and drive the CPU identity
through new mock hooks instead of mutating upstream package variables,
so they run everywhere. TestNotAMD64Host no longer depends on the host
architecture, and therefore now runs on amd64 too.
